Start rpc server in ObjectServer
